push: implement push options
https://git-scm.com/docs/git-push#Documentation/git-push.txt---push-optionltoptiongt
Push options are a way for a user to transmit strings to a git server via the command line. Push options are then passed to the `pre-receive` and `post-receive` hooks, where the server is able to interpret and make use of the user options. Some git repository hosting services like GitLab expose special options for users, like skipping CI tests on a specific push.

Co-Authored-By: pireads <pireads@gmail.com>
Co-Authored-By: lotdeef <lcfisch2@asu.edu>
Co-Authored-By: PSI497 <497.psi.497@gmail.com>